+++ src/SortableTree/components/TreeItemStructure/index.tsx
@@ -1,3 +1,5 @@
+import { TreeItem } from '../../types';
+
 type AriaProps = {
   'aria-label'?: string;
   'aria-labelledby'?: string;
@@ -5,6 +7,7 @@ type AriaProps = {
 };
 
 export interface TreeItemStructureProps {
+  treeItem: TreeItem;
   dropZoneRef: (element: HTMLElement | null) => void;
   draggableItemRef: React.Ref<any>;
   dropZoneStyle?: React.CSSProperties;
@@ -16,6 +19,7 @@ export interface TreeItemStructureProps {
   asDropZone?: React.ElementType;
   asDraggableItem?: React.ElementType;
   children?: React.ReactNode;
+  clone?: boolean;
   dataSlots: {
     dropZone?: AriaProps & Record<string, string | boolean | number | undefined>;
     draggableItem?: AriaProps & Record<string, string>;
@@ -32,14 +36,16 @@ export const TreeItemStructure = ({
   asDraggableItem: DraggableComponent = 'div',
   children,
   dataSlots,
+  treeItem,
+  clone,
 }: TreeItemStructureProps) => {
   return (
     <DropZoneComponent
+      data-tree-item-parent-id={treeItem.parentId}
       className={classNames.dropZone}
       style={dropZoneStyle}
-      {...dataSlots.dropZone}
+      {...(clone ? null : { ...dataSlots.dropZone, role: 'treeitem' })}
       ref={dropZoneRef}
-      role="treeitem"
     >
       <DraggableComponent
         className={classNames.draggableItem}
